Implement fine-grained locking.
This is in response to issue #50.  I have a partner who wants to make sure
that all calls form their "main" thread never stall waiting on the global
lock, even for a few ms, and has documented some times when some stalls
were visible due to contention.

Using tons of locks to cover lots of different objects is really scary.
A few high level strategies were used to make this possible:
- Document what locks protect what data.  Sprinkle asserts liberally
  when accessing that data to ensure that the appropriate lock is held.
- Restrict lock aquisition patterns so that deadlocks are impossible,
  and add logic to detect violations of the expected lock pattern.
- Use "leaf" locks smartyly.  A leaf lock is one that only protects
  a relatively small piece of data, and for which the usage pattern
  is "lock, access data quickly, unlock".  No other locks are allowed
  to be taken while a leaf lock is held, and so we don't have to worry
  about deadlocks.
- Major state changes, object creation and destruction, and other
  operations that are likely to touch multiple objects are still
  protected by the global lock.  Some API calls still take the global
  lock.  The most common API calls we expect to happen per frame and
  need to be fast, however, no longer take the global lock.
- Using scope lock objects, not only to take advantage of RAII unlocking,
  but also to clearly communicate lock ownership semantics between
  caller and calee.

Here are all the locks:
- Global lock.  All background processing and most "major" operations
  like creating and destroying objects, is done while holding this lock.
- Connection and poll group locks.  You must hold this lock while queuing
  messages or pulling them off the queue.
- Handle table lock.  The important API calls that need to be optimized
  begin by looking up a handle in a table and getting an object.  We have
  a lock to protect these tables.  (We use the same lock for both connection
  and pollgroup tables.)  The access pattern for this lock is the most
  unique -- see the code for details.
- Thinker lock.  Think times, and the PQ of thinkers, cannot be modified
  without holding this lock.  This lock also has a relatively unique access
  pattern.
- A single global recv message queue lock.  Received messages waiting to
  be fetched by the API are complicated because they may exist in multiple
  queues at the same time.  Instead of trying to have a lock for each queue
  and dealing with safely aquiring multiple locks when necessary, we just
  use a single "leaf" lock to protect *all* queues.  This is OK because
  the lock is only held very briefly.
- Misc other "leaf" locks.

If you are only using UDP connections, with each connection having its own
UDP socket, this may seem like overkill.  But the relationship between
connection and UDP socket is most definitely not one-to-one in the case of
SDR, where many connections utilize the same data structures.  One connection
may use one or more sessions to relays, and the server state data may be shared
between multiple connections.  These server data may use different UDP sockets,
or the sockets may be shared.  Basically each connection doesn't simply have
all its own data in such a way that it can be easily siloed.

There are still a few FIXMEs tagged in the code that I know need to be fixed,
but I confirmed basic functionality work in the simple UDP case, as well as
the much more complicated SDR case.  So this seems like a good time to check it
into a branch and get some broader testing on it.